# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Aerial Housing Detection API
  description: API para detecção de habitações em imagens aéreas com cruzamento CadUnico
  version: 1.0.0
  contact:
    name: Prefeitura de Paulista-PE
    email: ti@paulista.pe.gov.br
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8000
    description: Desenvolvimento local
  - url: https://staging.aerial.paulista.pe.gov.br
    description: Staging
  - url: https://aerial.paulista.pe.gov.br
    description: Producao
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      summary: Registrar novo usuario
      operationId: register
      tags: [auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email ja cadastrado
  /api/v1/auth/login:
    post:
      summary: Autenticar usuario
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login bem sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/refresh:
    post:
      summary: Renovar access token
      operationId: refresh
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token renovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/me:
    get:
      summary: Dados do usuario autenticado
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Dados do usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/auth/api-keys:
    post:
      summary: Criar API key
      operationId: createApiKey
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResponse"
    get:
      summary: Listar API keys do usuario
      operationId: listApiKeys
      tags: [auth]
      responses:
        "200":
          description: Lista de API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeyListItem"
  /api/v1/auth/api-keys/{key_id}:
    delete:
      summary: Revogar API key
      operationId: revokeApiKey
      tags: [auth]
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API key revogada
        "404":
          description: API key nao encontrada
  /api/v1/detect/analyze:
    post:
      summary: Analisar imagem aerea
      operationId: analyzeImage
      tags: [detect]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                confidence_threshold:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  default: 0.5
                region:
                  type: string
      responses:
        "200":
          description: Analise concluida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Imagem excede tamanho maximo
  /api/v1/detect/analyze/batch:
    post:
      summary: Analisar multiplas imagens
      operationId: analyzeBatch
      tags: [detect]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
                confidence_threshold:
                  type: number
                  default: 0.5
      responses:
        "200":
          description: Analise batch concluida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchAnalysisResponse"
  /api/v1/detect/results/{analysis_id}:
    get:
      summary: Consultar resultado de analise
      operationId: getResult
      tags: [detect]
      parameters:
        - name: analysis_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resultado da analise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResponse"
        "404":
          description: Analise nao encontrada
...
  /api/v1/lgpd/consent:
    post:
      summary: Registrar consentimento
      operationId: registerConsent
      tags: [lgpd]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentRequest"
      responses:
        "201":
          description: Consentimento registrado
  /api/v1/lgpd/consent/{subject_id}:
    get:
      summary: Consultar consentimentos de um titular
      operationId: getConsent
      tags: [lgpd]
      parameters:
        - name: subject_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Consentimentos do titular
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequest:
      description: Requisicao invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Nao autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Sem permissao
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Limite de requisicoes excedido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        role:
          type: string
          enum: [viewer, analyst, admin]
          default: viewer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        expires_in_days:
          type: integer
          default: 90
    ApiKeyResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
          description: Retornado apenas na criacao
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    ApiKeyListItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...
    AnalysisResponse:
      type: object
      properties:
        analysis_id:
          type: string
        status:
          type: string
          enum: [completed, failed]
        image_metadata:
          type: object
          properties:
            filename:
              type: string
            width:
              type: integer
            height:
              type: integer
            format:
              type: string
        detections:
          type: array
          items:
            $ref: "#/components/schemas/Detection"
        summary:
          type: object
          properties:
            total_detections:
              type: integer
            average_confidence:
              type: number
            processing_time_ms:
              type: number
        created_at:
          type: string
          format: date-time
    Detection:
      type: object
      properties:
        id:
          type: string
        bbox:
          $ref: "#/components/schemas/BBox"
        confidence:
          type: number
        class_name:
          type: string
        stage1_confidence:
          type: number
        stage2_confidence:
          type: number
    BBox:
      type: object
      properties:
        x_min:
          type: number
        y_min:
          type: number
        x_max:
          type: number
        y_max:
          type: number
    BatchAnalysisResponse:
      type: object
      properties:
        batch_id:
          type: string
        total_images:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisResponse"
        total_processing_time_ms:
          type: number
...
    ConsentRequest:
      type: object
      required: [subject_id, purpose]
      properties:
        subject_id:
          type: string
        purpose:
          type: string
        legal_basis:
          type: string
...